TOOLS-2974 Add tests for 5.1 & 5.2, fix multiple tests  (#412)
This PR includes multiple additions and fixes to testing:

- Fix some bugs in `binaryurl.py` that caused it not to find the correct URL.
- Set the project in the evergreen validate command (TOOLS-2515).
- Update merge key syntax in `common.yml` (TOOLS-2906).
- Put the shell commands for `"run qa-tests"` and `"run native-cert-ssl"` into their own bash scripts where we get the correct python binary, setup venv, and install python dependencies.
- Copy `netsnmp.dll` to the `bin` folder when downloading `mongod` on Windows.
- Add tests for 5.1 and 5.2 (TOOLS-2974 and TOOLS-2979).
- Download the enterprise edition for all distros where possible (AWS auth tests require the enterprise edition).
- Do not run AWS auth tests in macOS distro since it uses the community edition.
- Skip kerberos tests on SUSE distros (TODO: TOOLS-3014).
- Do not run MMAPv1 tests on PPC RHEL 7.1 (mmap is does not work for some reason. Didn't open a ticket for this since mmap is deprecated anyway).
- Skip AWS auth tests on PPC RHEL 8.1 (TODO: TOOLS-3015).
- Add correct resmoke SSL/TLS args to ZAP distros.
- Regenerate test TLS certs so the CA and IA have different subjects.
- Add the new format for build tags (https://go.dev/doc/go1.17#build-lines).
- Change the timing in `TestManagerStartAndStop()` to be less stringent so it is less flaky.
- Increase time in `TestBasicProgressBar()` so it is less flaky.
- Close open files in `TestTimeseriesCollections()`.
- Skip the `dropDatabase` check in `TestOplogRestoreUpdatesIndexCatalog()` (TODO: TOOLS-3016).
- Update unsupported commands in js tests.
- Convert resmoke from python2 to python3.
- Skip:
    - `dumping_dropped_collections.js` (TODO: TOOLS-3019)
    - `write_concern.js` (TODO: TOOLS-3017)
    - `write_concern_mongos.js` (TODO: TOOLS-3017)
    - `stat_discover.js` (TODO: TOOLS-3018)

+++ test/qa-tests/buildscripts/resmokelib/config.py
@@ -2,7 +2,7 @@
 Configuration options for resmoke.py.
 """
 
-from __future__ import absolute_import
+
 
 import os
 import os.path
@@ -52,7 +52,7 @@ DEFAULTS = {
     "nojournal": False,
     "repeat": 1,
     "reportFile": None,
-    "seed": long(time.time() * 256),  # Taken from random.py code in Python 2.7.
+    "seed": int(time.time() * 256),  # Taken from random.py code in Python 2.7.
     "shellReadMode": None,
     "shellWriteMode": None,
     "shuffle": False,
